scripts/fake_clippy/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fake-clippy"
version = "1.0.0"
description = "Brings back the 'helpful' Microsoft Office Assistant"
dependencies = [
    "pillow>=9.0.0"
]

[project.scripts]
fake-clippy = "fake_clippy:main"